Conflicts:
	Library/assetDatabase3
	Library/guidmapper

+++ Assets/Scripts/IAvatarBehaviour.cs
@@ -2,7 +2,7 @@
 
 public interface IAvatarBehaviour
 {
-    void Forward(int moveSpeed);
+    void Forward(float moveSpeed);
     void Left();
     void Right();
 }
